refactor(#101): Payment Entity 수정
- status 클래스명 Status로 수정
- Payment Entity 잘못된 import 수정

closed: #101
