+++ src/main/java/com/example/surveyapp/domain/point/domain/model/enums/Status.java
@@ -3,7 +3,7 @@ package com.example.surveyapp.domain.point.domain.model.enums;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
-public enum status {
+public enum Status {
     DONE("완료"), FAILED("실패");
 
     private final String status;
